seperate bucket from blockdb to bucket.go
