+++ src/visor/blockdb/blockdb.go
@@ -70,46 +70,6 @@ func Stop() {
 	db.Close()
 }
 
-type bucket struct {
-	Name []byte
-}
-
-func newBucket(name []byte) (*bucket, error) {
-	if !Disabled {
-		err := db.Update(func(tx *bolt.Tx) error {
-			if _, err := tx.CreateBucketIfNotExists(name); err != nil {
-				return err
-			}
-			return nil
-		})
-		if err != nil {
-			return nil, err
-		}
-	}
-	return &bucket{name}, nil
-}
-
-func (b bucket) Get(key []byte) []byte {
-	if Disabled {
-		return nil
-	}
-	var value []byte
-	db.View(func(tx *bolt.Tx) error {
-		value = tx.Bucket(b.Name).Get(key)
-		return nil
-	})
-	return value
-}
-
-func (b bucket) Set(key []byte, value []byte) error {
-	if Disabled {
-		return nil
-	}
-	return db.Update(func(tx *bolt.Tx) error {
-		return tx.Bucket(b.Name).Put(key, value)
-	})
-}
-
 // SetBlock save coin block.
 func SetBlock(block coin.Block) error {
 	//write to DB
